Bring all files in line with Google style. Fix some typos.

+++ dataEng_container_tools/gcs.py
@@ -5,25 +5,27 @@ from google.cloud import storage
 import pickle
 import os
 
+
 class gcs_file_io:
     """ """
     gcs_client = None
     gcs_secret_location = None
 
-    def __init__(self, gcs_secret_location, local = False):
+    def __init__(self, gcs_secret_location, local=False):
         self.gcs_secret_location = gcs_secret_location
         self.local = local
-        with open(gcs_secret_location,'r') as f:
+        with open(gcs_secret_location, 'r') as f:
             gcs_sa = json.load(f)
         with open('gcs-sa.json', 'w') as json_file:
             json.dump(gcs_sa, json_file)
-        self.gcs_client = storage.Client.from_service_account_json('gcs-sa.json')
+        self.gcs_client = storage.Client.from_service_account_json(
+            'gcs-sa.json')
 
     def __get_parts(self, gcs_uri):
         if gcs_uri.startswith('gs://'):
             gcs_uri = gcs_uri[5:]
         bucket = gcs_uri[:gcs_uri.find("/")]
-        file_path = gcs_uri[gcs_uri.find("/")+1:]
+        file_path = gcs_uri[gcs_uri.find("/") + 1:]
         return bucket, file_path
 
     def __wildcard_download(self, gcs_uri, default_file_type, dtype):
@@ -33,11 +35,15 @@ class gcs_file_io:
         prefix = file_path.strip('*')
         blobs = list(bucket.list_blobs(prefix=prefix))
         for blob in blobs:
-            return_dict[blob.name] = self.download_file_to_object(blob.name, default_file_type=default_file_type,
-                                                                  dtype=dtype)
+            return_dict[blob.name] = self.download_file_to_object(
+                blob.name, default_file_type=default_file_type, dtype=dtype)
         return return_dict
 
-    def download_file_to_object(self, gcs_uri, default_file_type = None, dtype = None, delimiter = None):
+    def download_file_to_object(self,
+                                gcs_uri,
+                                default_file_type=None,
+                                dtype=None,
+                                delimiter=None):
         """
 
         Args:
@@ -61,18 +67,34 @@ class gcs_file_io:
             bucket = self.gcs_client.bucket(bucket_name)
             binary_object = bucket.blob(file_path).download_as_string()
             file_like_object = io.BytesIO(binary_object)
-        hasEnding = file_path.endswith('.parquet') or file_path.endswith('.csv') or file_path.endswith('.pkl')
-        if file_path.endswith('.parquet') or ((not hasEnding) and (default_file_type == 'parquet')):
-            return pd.read_parquet(file_like_object, dtype = dtype) if dtype else pd.read_parquet(file_like_object)
-        if file_path.endswith('.csv') or ((not hasEnding) and (default_file_type == 'csv')):
-            return pd.read_csv(file_like_object, dtype = dtype, delimiter=delimiter) if dtype else pd.read_csv(file_like_object, delimiter=delimiter)
-        if file_path.endswith('.pkl') or ((not hasEnding) and (default_file_type == 'pkl')):
-            return pd.read_pickle(file_like_object) if dtype else pd.read_pickle(file_like_object)#, dtype = dtype
-        if file_path.endswith('.json') or ((not hasEnding) and (default_file_type == 'json')):
+        hasEnding = file_path.endswith('.parquet') or file_path.endswith(
+            '.csv') or file_path.endswith('.pkl')
+        if file_path.endswith('.parquet') or ((not hasEnding) and
+                                              (default_file_type == 'parquet')):
+            return pd.read_parquet(
+                file_like_object,
+                dtype=dtype) if dtype else pd.read_parquet(file_like_object)
+        if file_path.endswith('.csv') or ((not hasEnding) and
+                                          (default_file_type == 'csv')):
+            return pd.read_csv(file_like_object,
+                               dtype=dtype,
+                               delimiter=delimiter) if dtype else pd.read_csv(
+                                   file_like_object, delimiter=delimiter)
+        if file_path.endswith('.pkl') or ((not hasEnding) and
+                                          (default_file_type == 'pkl')):
+            return pd.read_pickle(
+                file_like_object) if dtype else pd.read_pickle(
+                    file_like_object)  #, dtype = dtype
+        if file_path.endswith('.json') or ((not hasEnding) and
+                                           (default_file_type == 'json')):
             return json.load(file_like_object)
         return file_like_object
 
-    def download_files_to_objects(self, gcs_uris, default_file_type = None, dtypes = [], delimiters = []):
+    def download_files_to_objects(self,
+                                  gcs_uris,
+                                  default_file_type=None,
+                                  dtypes=[],
+                                  delimiters=[]):
         """
 
         Args:
@@ -96,11 +118,15 @@ class gcs_file_io:
                 delim = delimiters[0]
             elif len(delimiters) > 1:
                 delim = delimiters[pos]
-            return_objects.append(self.download_file_to_object(gcs_uri, default_file_type=default_file_type,
-                                  dtype = dt, delimiter = delim))
+            return_objects.append(
+                self.download_file_to_object(
+                    gcs_uri,
+                    default_file_type=default_file_type,
+                    dtype=dt,
+                    delimiter=delim))
         return return_objects
 
-    def download_file_to_disk(self, gcs_uri, local_location = None):
+    def download_file_to_disk(self, gcs_uri, local_location=None):
         """
 
         Args:
@@ -122,7 +148,7 @@ class gcs_file_io:
         bucket.get_blob(file_path).download_to_filename(local_location)
         return local_location
 
-    def download_files_to_disk(self, gcs_uris, local_locations = []):
+    def download_files_to_disk(self, gcs_uris, local_locations=[]):
         """
 
         Args:
@@ -135,12 +161,15 @@ class gcs_file_io:
         return_locations = []
         for pos, gcs_uri in enumerate(gcs_uris):
             if len(local_locations) > 0:
-                return_locations.append(self.download_file_to_disk(gcs_uri = gcs_uri, local_location = local_locations[pos]))
+                return_locations.append(
+                    self.download_file_to_disk(
+                        gcs_uri=gcs_uri, local_location=local_locations[pos]))
             else:
-                return_locations.append(self.download_file_to_disk(gcs_uri = gcs_uri))
+                return_locations.append(
+                    self.download_file_to_disk(gcs_uri=gcs_uri))
         return return_locations
 
-    def upload_file_from_disk(self, gcs_uri, local_location, metadata = {}):
+    def upload_file_from_disk(self, gcs_uri, local_location, metadata={}):
         """
 
         Args:
@@ -184,15 +213,26 @@ class gcs_file_io:
         """
         return_objects = []
         for pos, gcs_uri in enumerate(gcs_uris):
-            if len(metadata) ==0:
-                return_objects.append(self.upload_file_from_disk(gcs_uri, local_locations[pos]))
+            if len(metadata) == 0:
+                return_objects.append(
+                    self.upload_file_from_disk(gcs_uri, local_locations[pos]))
             elif len(metadata) == 1:
-                return_objects.append(self.upload_file_from_disk(gcs_uri, local_locations[pos], metadata=metadata[0]))
+                return_objects.append(
+                    self.upload_file_from_disk(gcs_uri,
+                                               local_locations[pos],
+                                               metadata=metadata[0]))
             else:
-                return_objects.append(self.upload_file_from_disk(gcs_uri, local_locations[pos], metadata=metadata[pos]))
+                return_objects.append(
+                    self.upload_file_from_disk(gcs_uri,
+                                               local_locations[pos],
+                                               metadata=metadata[pos]))
         return return_objects
 
-    def upload_file_from_object(self, gcs_uri, object_to_upload, default_file_type = None, metadata = {}):
+    def upload_file_from_object(self,
+                                gcs_uri,
+                                object_to_upload,
+                                default_file_type=None,
+                                metadata={}):
         """
 
         Args:
@@ -221,37 +261,46 @@ class gcs_file_io:
             bucket = self.gcs_client.bucket(bucket_name)
             blob = bucket.blob(file_path)
             blob.metadata = metadata
-        hasEnding = file_path.endswith('.parquet') or file_path.endswith('.csv') or file_path.endswith('.pkl')
-        if file_path.endswith('.parquet') or ((not hasEnding) and (default_file_type == 'parquet')):
+        hasEnding = file_path.endswith('.parquet') or file_path.endswith(
+            '.csv') or file_path.endswith('.pkl')
+        if file_path.endswith('.parquet') or ((not hasEnding) and
+                                              (default_file_type == 'parquet')):
             if self.local:
                 return object_to_upload.to_parquet(gcs_uri)
             fileObject = io.BytesIO()
             object_to_upload.to_parquet(fileObject)
             fileObject.seek(0)
             return blob.upload_from_file(fileObject)
-        if file_path.endswith('.csv') or ((not hasEnding) and (default_file_type == 'csv')):
+        if file_path.endswith('.csv') or ((not hasEnding) and
+                                          (default_file_type == 'csv')):
             if self.local:
                 return object_to_upload.to_csv(gcs_uri)
             csv_string = object_to_upload.to_csv(encoding='utf-8')
             return blob.upload_from_string(csv_string)
-        if file_path.endswith('.pkl') or ((not hasEnding) and (default_file_type == 'pkl')):
+        if file_path.endswith('.pkl') or ((not hasEnding) and
+                                          (default_file_type == 'pkl')):
             if self.local:
-                return pickle.dump(gcs_uri, open(gcs_uri,'wb'))
+                return pickle.dump(gcs_uri, open(gcs_uri, 'wb'))
             fileObject = io.BytesIO(pickle.dumps(object_to_upload))
             fileObject.seek(0)
             return blob.upload_from_file(fileObject)
-        if file_path.endswith('.json') or ((not hasEnding) and (default_file_type == 'json')):
+        if file_path.endswith('.json') or ((not hasEnding) and
+                                           (default_file_type == 'json')):
             if self.local:
-                return json.dump(gcs_uri, open(gcs_uri,'w'))
+                return json.dump(gcs_uri, open(gcs_uri, 'w'))
             json_string = json.dumps(fileObject)
             return blob.upload_from_string(json_string)
         if self.local:
-            return pickle.dump(gcs_uri, open(gcs_uri,'wb'))
+            return pickle.dump(gcs_uri, open(gcs_uri, 'wb'))
         fileObject = io.BytesIO(pickle.dumps(object_to_upload))
         fileObject.seek(0)
         return blob.upload_from_file(fileObject)
 
-    def upload_files_from_objects(self, gcs_uris, objects_to_upload, default_file_type = None, metadata = []):
+    def upload_files_from_objects(self,
+                                  gcs_uris,
+                                  objects_to_upload,
+                                  default_file_type=None,
+                                  metadata=[]):
         """
 
         Args:
@@ -265,10 +314,24 @@ class gcs_file_io:
         """
         return_objects = []
         for pos, gcs_uri in enumerate(gcs_uris):
-            if len(metadata)==0:
-                 return_objects.append(self.upload_file_from_object(gcs_uri, objects_to_upload[pos], default_file_type = default_file_type))
+            if len(metadata) == 0:
+                return_objects.append(
+                    self.upload_file_from_object(
+                        gcs_uri,
+                        objects_to_upload[pos],
+                        default_file_type=default_file_type))
             elif len(metadata) == 1:
-                return_objects.append(self.upload_file_from_object(gcs_uri, objects_to_upload[pos], default_file_type = default_file_type, metadata=metadata[0]))
+                return_objects.append(
+                    self.upload_file_from_object(
+                        gcs_uri,
+                        objects_to_upload[pos],
+                        default_file_type=default_file_type,
+                        metadata=metadata[0]))
             else:
-                return_objects.append(self.upload_file_from_object(gcs_uri, objects_to_upload[pos], default_file_type = default_file_type, metadata=metadata[pos]))
+                return_objects.append(
+                    self.upload_file_from_object(
+                        gcs_uri,
+                        objects_to_upload[pos],
+                        default_file_type=default_file_type,
+                        metadata=metadata[pos]))
         return return_objects
